Support lint on save

+++ lib/Model/Linter/PhpstanLinter.php
@@ -3,8 +3,11 @@
 namespace Phpactor\Extension\LanguageServerPhpstan\Model\Linter;
 
 use Amp\Promise;
+use Generator;
+use LanguageServerProtocol\Diagnostic;
 use Phpactor\Extension\LanguageServerPhpstan\Model\Linter;
 use Phpactor\Extension\LanguageServerPhpstan\Model\PhpstanProcess;
+use Phpactor\TextDocument\TextDocumentUri;
 use function Safe\tempnam;
 use function Safe\file_put_contents;
 
@@ -24,12 +27,25 @@ class PhpstanLinter implements Linter
     {
         /** @phpstan-ignore-next-line */
         return \Amp\call(function () use ($url, $text) {
-            $name = tempnam(sys_get_temp_dir(), 'phpstanls');
-            file_put_contents($name, $text);
-            $diagnostics = yield $this->process->analyse($name);
-            unlink($name);
+            $diagnostics = yield from $this->doLint($url, $text);
 
             return $diagnostics;
         });
+    }
+
+    /**
+     * @return Generator<Promise<array<Diagnostic>>>
+     */
+    private function doLint(string $url, ?string $text): Generator
+    {
+        if (null === $text) {
+            return yield $this->process->analyse(TextDocumentUri::fromString($url)->path());
+        }
+
+        $name = tempnam(sys_get_temp_dir(), 'phpstanls');
+        file_put_contents($name, $text);
+        $diagnostics = yield $this->process->analyse($name);
+        unlink($name);
+        return $diagnostics;
     }
 }
